Separate TagVerifierTest from ContentFilterTest
Move TagVerifierTest out of ContentFilterTest into its own file. (Why
was it nested anyway? Just to share some const fields?)
